src/requirements.txt
astroid==2.2.5
autopep8==1.4.4
awscli==1.16.202
bcrypt==3.1.7
boto3==1.9.192
botocore==1.12.192
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
docutils==0.14
ecs-deploy==1.9.0
future==0.17.1
idna==2.8
imagecodecs==2020.2.18
imageio==2.8.0
isort==4.3.18
jmespath==0.9.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.0
matplotlib==3.2.1
mccabe==0.6.1
mysql-connector==2.2.9
networkx==2.4
numpy==1.18.4
pandas==1.0.3
paramiko==2.7.1
Pillow==7.1.2
powerline-gitstatus==1.3.1
powerline-status==2.7
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.3.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
sshtunnel==0.1.5
tifffile==2020.5.11
typed-ast==1.3.5
tzlocal==2.1
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.1
